move exception (#5041)
