+++ firedrake/__init__.py
@@ -70,10 +70,10 @@ from firedrake.deflation import DeflatedSNES, Deflation  # noqa: F401
 from firedrake.exceptions import (  # noqa: F401
     FiredrakeException, ConvergenceError, MismatchingDomainError,
     VertexOnlyMeshMissingPointsError, DofNotDefinedError, DofTypeError,
+    PointNotInDomainError,
 )
 from firedrake.function import (  # noqa: F401
-    Function, PointNotInDomainError,
-    CoordinatelessFunction, PointEvaluator
+    Function, CoordinatelessFunction, PointEvaluator
 )
 from firedrake.functionspace import (  # noqa: F401
     MixedFunctionSpace, FunctionSpace, VectorFunctionSpace,
